refactor: up required Node version to 22.14.0 and refactor imports to follow barrel pattern
