+++ package.json
@@ -1,23 +1,22 @@
 {
-    "name": "node-file-manager",
-    "version": "1.0.0",
-    "description": "This repository is the solution of nodejs-assignments https://github.com/AlreadyBored/nodejs-assignments/blob/main/assignments/file-manager/assignment.md",
-    "engines": {
-      "node": ">=22.0.9"
-    },
-    "type": "module",
-    "scripts": {
-      "start": "node src/app.js"
-    },
-    "repository": {
-      "type": "git",
-      "url": "git+https://github.com/zenitfan19/file-manager.git"
-    },
-    "keywords": [
-      "nodejs",
-      "assignments"
-    ],
-    "author": "zenitfan19",
-    "license": "ISC"
-  }
-  
+  "name": "node-file-manager",
+  "version": "1.0.0",
+  "description": "This repository is the solution of nodejs-assignments https://github.com/AlreadyBored/nodejs-assignments/blob/main/assignments/file-manager/assignment.md",
+  "engines": {
+    "node": ">=22.14.0"
+  },
+  "type": "module",
+  "scripts": {
+    "start": "node src/app.js"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/zenitfan19/file-manager.git"
+  },
+  "keywords": [
+    "nodejs",
+    "assignments"
+  ],
+  "author": "zenitfan19",
+  "license": "ISC"
+}
