+++ src/app.js
@@ -1,11 +1,12 @@
 import { stdin, stdout } from "node:process";
 import { homedir } from "node:os";
 import { createInterface } from "node:readline/promises";
-import { getCLIArgument } from "./utils/getCLIArgument.js";
-import { commandHandler } from "./utils/commandHandler.js";
-import { exitProcess } from "./services/exitProcess.js";
-import { printCurrentDirectory } from "./utils/printCurrentDirectory.js";
-import { changeDirectory } from "./services/changeDirectory.js";
+import {
+  getCLIArgument,
+  commandHandler,
+  printCurrentDirectory,
+} from "./utils/index.js";
+import { exitProcess, changeDirectory } from "./services/index.js";
 
 const userName = getCLIArgument("username");
