+++ src/services/addFile.js
@@ -1,7 +1,7 @@
 import { cwd } from "node:process";
 import { writeFile } from "node:fs/promises";
 import { join } from "node:path";
-import { checkFileAlreadyExists } from '../utils/checkFileAlreadyExists.js';
+import { checkFileAlreadyExists } from "../utils/index.js";
 
 const addFile = async (command) => {
   const fileName = command.slice(3).trim();
