+++ src/services/index.js
@@ -1,0 +1,13 @@
+export * from "./addFile.js";
+export * from "./calcHash.js";
+export * from "./changeDirectory.js";
+export * from "./compressFile.js";
+export * from "./copyFile.js";
+export * from "./decompressFile.js";
+export * from "./deleteFile.js";
+export * from "./exitProcess.js";
+export * from "./listDirectoryContent.js";
+export * from "./moveFile.js";
+export * from "./os.js";
+export * from "./readFile.js";
+export * from "./renameFile.js";
